requirements.txt
line-bot-sdk==1.18.0
flask==1.1.2